Add the begining of the proof for garble1
Most of the game for the case of one gate are written. It lacks maybe
one or two games but it's not sure. It lacks also the description
of garble function for one gate(and for a circuit btw) but except that
most is written and it remains to write some intermediate lemmas and prove
them.